+++ src/js/components/modals/__tests__/InstallPackageModal-test.js
@@ -8,9 +8,9 @@ jest.dontMock('../../../stores/CosmosPackagesStore');
 jest.dontMock('../../../mixins/InternalStorageMixin');
 jest.dontMock('../../../mixins/TabsMixin');
 jest.dontMock('../../../utils/FormUtil');
-jest.dontMock('../../../utils/GeminiUtil');
 jest.dontMock('../../../utils/SchemaFormUtil');
 jest.dontMock('../../../utils/SchemaUtil');
+jest.dontMock('../../../utils/ScrollbarUtil');
 jest.dontMock('../../../utils/Util');
 
 /* eslint-disable no-unused-vars */
